Update SOI reading code.
A recent update to the soiplaintext.html file on the Bureau's FTP server
has changed the data from being formatted purely with tabs. The most
recent entry is separated by spaces instead of a tab. This update
changes to use a whitespace matching regular expression in the pandas
read method.